Add code fixer for tab order
Handle scenarios without docking, e.g. all controls that have misaligned
`TabIndex` vs z-order aren't docked.

There are three distinct scenarios when it comes to re-ordering statements:

1. Move down by one, or swap neighbouring statements - in this case we
need to swap the leading trivia before we switch.

2. Move down by more than one position - we need to swap leading trivia
with the statement directly below the statement we are moving, and update
the trivia for the moved statement with the leading trivia of the statement
above which we're inserting.

3. Move up - we need to swap the leading trivia for the moved statement
with the leading trivia of the statement above which we're inserting, and
update that statement's trivia with the leading trivia of the statement
directly below it.